% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_get_dataflow_endpoint_group}
\alias{groundstation_get_dataflow_endpoint_group}
\title{Returns the dataflow endpoint group}
\usage{
groundstation_get_dataflow_endpoint_group(dataflowEndpointGroupId)
}
\arguments{
\item{dataflowEndpointGroupId}{[required] UUID of a dataflow endpoint group.}
}
\description{
Returns the dataflow endpoint group.
}
\section{Request syntax}{
\preformatted{svc$get_dataflow_endpoint_group(
  dataflowEndpointGroupId = "string"
)
}
}

\keyword{internal}
