% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundstation_operations.R
\name{groundstation_get_satellite}
\alias{groundstation_get_satellite}
\title{Returns a satellite}
\usage{
groundstation_get_satellite(satelliteId)
}
\arguments{
\item{satelliteId}{[required] UUID of a satellite.}
}
\description{
Returns a satellite.
}
\section{Request syntax}{
\preformatted{svc$get_satellite(
  satelliteId = "string"
)
}
}

\keyword{internal}
